\name{yaqc-methods}
\docType{methods}
\alias{yaqc-methods}
\alias{yaqc,eSet-method}

\title{Create an YAQCStat object}

\description{ Create an YAQCStats qc object for an AffyBatch (or
  ExpressionSet) input } 

\section{Methods}{
  \describe{
    \item{object="eSet"}{ Create a full YAQCStats object for an AffyBatch
      input or a light YAQCStats object for an ExpressionSet input. } 
  }
}

\seealso{ \code{\link{yaqc.affy}}}

\keyword{methods}

